requirements.txt
aiodns==1.1.1
aiohttp==2.0.3
aiohttp-jinja2==0.13.0
async-timeout==1.2.0
cchardet==1.1.3
Cerberus==1.1
motor==1.1
PyYAML==3.12
uvloop==0.8.0
